seperated icons to icons.js
